moved IniReaderTrait into Feature subdir
