+++ module/VuFind/src/VuFind/Config/PathResolverFactory.php
@@ -35,6 +35,7 @@ use Laminas\ServiceManager\Exception\ServiceNotFoundException;
 use Laminas\ServiceManager\Factory\FactoryInterface;
 use Psr\Container\ContainerExceptionInterface as ContainerException;
 use Psr\Container\ContainerInterface;
+use VuFind\Config\Feature\IniReaderTrait;
 
 /**
  * Factory for PathResolver.
